close #3
This commit adds support for self-signed certs at the splunk endoint.
Users can (optionally) add a set of CAs that the logger should trust
when hitting the splunk endpoint, and if the logger sees CAs specified
in the config file it will ignore the system trust store and only
validate certs against the configured CAs.